error arreglat

+++ src/persistencia/GestorExpBooleanes.java
@@ -18,17 +18,19 @@ public class GestorExpBooleanes {
             File carpeta = new File(dirPath);
             if (carpeta.exists()) {
                 File[] llistaFitxers = carpeta.listFiles();
-                for (File fitxer : llistaFitxers) {
-                    String nom = fitxer.getName();
-                    FileInputStream fileInputStream = new FileInputStream(dirPath.concat(nom));
-                    ObjectInputStream objectInputStream = new ObjectInputStream(fileInputStream);
-                    String exp = (String) objectInputStream.readObject();
-                    objectInputStream.close();
-                    Pair<String, String> p = new Pair<>();
-                    p.x = nom;
-                    p.y = exp;
-                    exps.add(p);
-                    fitxer.delete();
+                if (llistaFitxers != null) {
+                    for (File fitxer : llistaFitxers) {
+                        String nom = fitxer.getName();
+                        FileInputStream fileInputStream = new FileInputStream(dirPath.concat(nom));
+                        ObjectInputStream objectInputStream = new ObjectInputStream(fileInputStream);
+                        String exp = (String) objectInputStream.readObject();
+                        objectInputStream.close();
+                        Pair<String, String> p = new Pair<>();
+                        p.x = nom;
+                        p.y = exp;
+                        exps.add(p);
+                        fitxer.delete();
+                    }
                 }
             }
             return exps;
@@ -42,13 +44,13 @@ public class GestorExpBooleanes {
    public static void GuardarExpB(List<Pair<String, String>> exps) throws IOException {
         String dirPath = "./appdata/expressions/";
         Files.createDirectories(Paths.get(dirPath));
-        for (int i = 0; i < exps.size(); ++i) {
-            FileOutputStream fileOutputStream = new FileOutputStream(dirPath.concat(exps.get(i).x));
-            ObjectOutputStream objectOutputStream = new ObjectOutputStream(fileOutputStream);
-            objectOutputStream.writeObject(exps.get(i).y);
-            objectOutputStream.flush();
-            objectOutputStream.close();
-        }
+       for (Pair<String, String> exp : exps) {
+           FileOutputStream fileOutputStream = new FileOutputStream(dirPath.concat(exp.x));
+           ObjectOutputStream objectOutputStream = new ObjectOutputStream(fileOutputStream);
+           objectOutputStream.writeObject(exp.y);
+           objectOutputStream.flush();
+           objectOutputStream.close();
+       }
     }
 }
 /*    public static void GuardarExpB(List<Pair<String, String>> exps) throws IOException {
